requirements.txt
Django==4.1.1
psycopg2==2.9.3
djangorestframework==3.13.1
gunicorn==20.1.0 
Pillow==9.1.1
channels==3.0.4
channels-redis==3.3.1
requests==2.28.1
dj-rest-auth==2.2.5
django-allauth==0.51.0
django-cors-headers==4.2.0
newsapi-python==0.2.7
